Relocate time provider classes to domain module
