+++ domain/src/test/java/akio/apps/myrun/domain/tracking/AverageLocationAccumulatorTest.kt
@@ -2,7 +2,7 @@ package akio.apps.myrun.domain.tracking
 
 import akio.apps.myrun.data.location.api.model.Location
 import akio.apps.myrun.domain.tracking.locationprocessor.AverageLocationAccumulator
-import akio.apps.myrun.wiring.common.TimeProvider
+import akio.apps.myrun.domain.time.TimeProvider
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.Assert.assertEquals
 import org.junit.Before
