+++ feature/home/src/main/java/akio/apps/myrun/feature/home/feed/ActivityPagingSource.kt
@@ -3,7 +3,6 @@ package akio.apps.myrun.feature.home.feed
 import akio.apps.myrun.data.activity.api.model.BaseActivityModel
 import akio.apps.myrun.data.common.Resource
 import akio.apps.myrun.domain.activity.GetFeedActivitiesUsecase
-import akio.apps.myrun.wiring.common.Now
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
 import javax.inject.Inject
@@ -13,7 +12,7 @@ class ActivityPagingSource @Inject constructor(
     private val getFeedActivitiesUsecase: GetFeedActivitiesUsecase,
 ) : PagingSource<Long, BaseActivityModel>() {
     override suspend fun load(params: LoadParams<Long>): LoadResult<Long, BaseActivityModel> {
-        val startAfter = params.key ?: Now.currentTimeMillis()
+        val startAfter = params.key ?: System.currentTimeMillis()
         val resource = getFeedActivitiesUsecase.getUserTimelineActivity(startAfter, params.loadSize)
         Timber.d("feed resource paramKey=${params.key} startAfter=$startAfter")
         return when (resource) {
